pyproject.toml
[project]
name = "v2-partitioner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.72",
    "chonkie[all]>=1.4.1",
    "cuid2>=2.0.1",
    "datalab-python-sdk>=0.1.13",
    "fastapi[standard]>=0.121.1",
    "firecrawl-py>=4.8.0",
    "markitdown[docx,outlook,pptx,xls,xlsx]>=0.1.3",
    "modal>=1.2.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "python-magic>=0.4.27",
    "redis>=7.0.1",
    "youtube-transcript-api>=1.2.3",
]
